[#1153] option D * Updating NEWS and release notes to current changes 	modified:   NEWS 	modified:   doc/src/release_notes.rst * Adjusting documentation 	modified:   doc/queries/doc-pgr_TSPannealing.queries 	modified:   doc/tsp/TSP-family.rst 	modified:   doc/tsp/pgr_TSPannealing.rst * Renaming files 	renamed:    include/drivers/tsp/euclideanTSP_driver.h -> include/drivers/tsp/TSPannealingEuclidean_driver.h 	renamed:    include/drivers/tsp/newTSP_driver.h -> include/drivers/tsp/TSPannealing_driver.h 	renamed:    pgtap/tsp/TSPeuclidean/illegal_values.sql -> pgtap/tsp/TSPannealing/coordinates-illegal_values.sql 	renamed:    pgtap/tsp/TSPeuclidean/no_crash_test.sql -> pgtap/tsp/TSPannealing/coordinates-no_crash_test.sql 	renamed:    pgtap/tsp/TSPeuclidean/innerQuery.sql -> pgtap/tsp/TSPannealing/innerQuery.sql 	renamed:    pgtap/tsp/pgr_TSP_types_check.sql -> pgtap/tsp/TSPannealing/matrix-illegal_values.sql 	renamed:    pgtap/tsp/no_crash_test-tsp.sql -> pgtap/tsp/TSPannealing/matrix-no_crash_test.sql 	renamed:    pgtap/tsp/TSPeuclidean/types_check.sql -> pgtap/tsp/TSPannealing/types_check.sql 	renamed:    src/tsp/euclideanTSP.c -> src/tsp/TSPannealing.c 	renamed:    src/tsp/euclideanTSP_driver.cpp -> src/tsp/TSPannealingEuclidean_driver.cpp 	renamed:    src/tsp/TSP_driver.cpp -> src/tsp/TSPannealing_driver.cpp * Updating signatures 	modified:   sql/sigs/pgrouting--3.0.0.sig * Function name pgr_TSPannealing 	modified:   sql/tsp/TSPannealing.sql 	modified:   sql/tsp/_TSPannealing.sql * Adjustment for reading both SQL 	modified:   src/common/coordinates_input.c 	modified:   src/common/matrixRows_input.c * Documentation queries adjustment 	modified:   test/tsp/doc-pgr_TSPannealing.result 	modified:   test/tsp/doc-pgr_TSPannealing.test.sql 	modified:   test/tsp/test.conf * Deleting unusedd files 	deleted:    doc/queries/doc-pgr_TSPannealingEuclidean.queries 	deleted:    doc/tsp/pgr_TSPannealingEuclidean.rst 	deleted:    sql/tsp/TSPannealingEuclidean.sql 	deleted:    sql/tsp/_TSPannealingEuclidean.sql 	deleted:    src/tsp/TSP.c * Adjusting build to current files 	modified:   doc/tsp/CMakeLists.txt 	modified:   doc/queries/CMakeLists.txt 	modified:   sql/tsp/CMakeLists.txt 	modified:   src/tsp/CMakeLists.txt
