+++ NEWS
@@ -5,7 +5,7 @@ pgRouting 3.0.0 Release Notes
 *Fixed Issues*
 
 * [#1153 ](https://github.com/pgRouting/pgrouting/issues/1153)_:
-  * Renamed pgr_eucledianTSP to pgr_TSPannealingEuclidean
+  * Renamed pgr_eucledianTSP to pgr_TSPannealing
   * Renamed pgr_TSP to pgr_TSPannealing
 
 *New Experimental functions*
