--- doc/queries/doc-pgr_TSPannealingEuclidean.queries
@@ -1,105 +1,0 @@
-BEGIN;
-BEGIN
-SET client_min_messages TO NOTICE;
-SET
--- q1
-SELECT * FROM pgr_TSPannealingEuclidean(
-    $$
-    SELECT id, st_X(the_geom) AS x, st_Y(the_geom)AS y  FROM edge_table_vertices_pgr
-    $$,
-    randomize := false);
- seq | node |       cost        |     agg_cost
------+------+-------------------+------------------
-   1 |    1 |   1.4142135623731 |                0
-   2 |    3 |                 1 |  1.4142135623731
-   3 |    4 |                 1 | 2.41421356237309
-   4 |    9 |                 1 | 3.41421356237309
-   5 |    6 |  0.58309518948453 | 4.41421356237309
-   6 |   16 | 0.860232526704263 | 4.99730875185763
-   7 |   12 |  1.11803398874989 | 5.85754127856189
-   8 |   17 |  1.11803398874989 | 6.97557526731178
-   9 |   11 |                 1 | 8.09360925606168
-  10 |   10 |               0.5 | 9.09360925606168
-  11 |   15 |               0.5 | 9.59360925606168
-  12 |   13 |  1.58113883008419 | 10.0936092560617
-  13 |   14 |  1.58113883008419 | 11.6747480861459
-  14 |    7 |                 1 | 13.2558869162301
-  15 |    8 |                 1 | 14.2558869162301
-  16 |    5 |                 1 | 15.2558869162301
-  17 |    2 |                 1 | 16.2558869162301
-  18 |    1 |                 0 | 17.2558869162301
-(18 rows)
-
--- q2
-SELECT* from pgr_TSPannealingEuclidean(
-    $$
-    SELECT id, st_X(the_geom) AS x, st_Y(the_geom) AS y FROM edge_table_vertices_pgr
-    $$,
-    tries_per_temperature := 3,
-    cooling_factor := 0.5,
-    randomize := false);
- seq | node |       cost       |     agg_cost
------+------+------------------+------------------
-   1 |    1 |  1.4142135623731 |                0
-   2 |    3 |                1 |  1.4142135623731
-   3 |    4 |                1 | 2.41421356237309
-   4 |    9 | 0.58309518948453 | 3.41421356237309
-   5 |   16 | 0.58309518948453 | 3.99730875185762
-   6 |    6 |                1 | 4.58040394134215
-   7 |    5 |                1 | 5.58040394134215
-   8 |    8 |                1 | 6.58040394134215
-   9 |    7 | 1.58113883008419 | 7.58040394134215
-  10 |   14 |   1.499999999999 | 9.16154277142634
-  11 |   15 |              0.5 | 10.6615427714253
-  12 |   13 |              1.5 | 11.1615427714253
-  13 |   17 | 1.11803398874989 | 12.6615427714253
-  14 |   12 |                1 | 13.7795767601752
-  15 |   11 |                1 | 14.7795767601752
-  16 |   10 |                2 | 15.7795767601752
-  17 |    2 |                1 | 17.7795767601752
-  18 |    1 |                0 | 18.7795767601752
-(18 rows)
-
--- q3
-SET client_min_messages TO DEBUG1;
-SET
-SELECT* from pgr_TSPannealingEuclidean(
-    $$
-    SELECT id, st_X(the_geom) AS x, st_Y(the_geom) AS y FROM edge_table_vertices_pgr
-    $$,
-    tries_per_temperature := 0,
-    randomize := false);
-DEBUG:  Processing Information
-Initializing tsp class ---> tsp.greedyInitial ---> tsp.annealing ---> OK
-
-Cycle(100) 	total changes =0	0 were because  delta energy < 0
-Total swaps: 3
-Total slides: 0
-Total reverses: 0
-Times best tour changed: 4
-Best cost reached = 18.7796
- seq | node |       cost       |     agg_cost
------+------+------------------+------------------
-   1 |    1 |  1.4142135623731 |                0
-   2 |    3 |                1 |  1.4142135623731
-   3 |    4 |                1 | 2.41421356237309
-   4 |    9 | 0.58309518948453 | 3.41421356237309
-   5 |   16 | 0.58309518948453 | 3.99730875185762
-   6 |    6 |                1 | 4.58040394134215
-   7 |    5 |                1 | 5.58040394134215
-   8 |    8 |                1 | 6.58040394134215
-   9 |    7 | 1.58113883008419 | 7.58040394134215
-  10 |   14 |   1.499999999999 | 9.16154277142634
-  11 |   15 |              0.5 | 10.6615427714253
-  12 |   13 |              1.5 | 11.1615427714253
-  13 |   17 | 1.11803398874989 | 12.6615427714253
-  14 |   12 |                1 | 13.7795767601752
-  15 |   11 |                1 | 14.7795767601752
-  16 |   10 |                2 | 15.7795767601752
-  17 |    2 |                1 | 17.7795767601752
-  18 |    1 |                0 | 18.7795767601752
-(18 rows)
-
--- q4
-ROLLBACK;
-ROLLBACK
