+++ sql/sigs/pgrouting--3.0.0.sig
@@ -188,8 +188,6 @@ _pgr_trsp(text,text,bigint,bigint,boolean)
 pgr_trspviaedges(text,integer[],double precision[],boolean,boolean,text)
 pgr_trspviavertices(text,anyarray,boolean,boolean,text)
 _pgr_trspviavertices(text,integer[],boolean,boolean,text)
-_pgr_tspannealingeuclidean(text,bigint,bigint,double precision,integer,integer,integer,double precision,double precision,double precision,boolean)
-pgr_tspannealingeuclidean(text,bigint,bigint,double precision,integer,integer,integer,double precision,double precision,double precision,boolean)
 _pgr_tspannealing(text,bigint,bigint,double precision,integer,integer,integer,double precision,double precision,double precision,boolean)
 pgr_tspannealing(text,bigint,bigint,double precision,integer,integer,integer,double precision,double precision,double precision,boolean)
 _pgr_turnrestrictedpath(text,text,bigint,bigint,integer,boolean,boolean,boolean,boolean)
