--- sql/tsp/_TSPannealingEuclidean.sql
@@ -1,56 +1,0 @@
-/*PGR-GNU*****************************************************************
-
-Copyright (c) 2016 pgRouting developers
-Mail: project@pgrouting.org
-
-Copyright (c) 2016 Celia Virginia Vergara Castillo
-mail: vicky_vergara@hotmail.com
-
-------
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-
-********************************************************************PGR-GNU*/
-
-
-CREATE OR REPLACE FUNCTION _pgr_TSPAnnealingEuclidean(
-    coordinates_sql TEXT,
-    start_id BIGINT DEFAULT 0,
-    end_id BIGINT DEFAULT 0,
-
-    max_processing_time FLOAT DEFAULT '+infinity'::FLOAT,
-
-    tries_per_temperature INTEGER DEFAULT 500,
-    max_changes_per_temperature INTEGER DEFAULT 60,
-    max_consecutive_non_changes INTEGER DEFAULT 100,
-
-    initial_temperature FLOAT DEFAULT 100,
-    final_temperature FLOAT DEFAULT 0.1,
-    cooling_factor FLOAT DEFAULT 0.9,
-
-    randomize BOOLEAN DEFAULT true,
-
-    OUT seq integer,
-    OUT node BIGINT,
-    OUT cost FLOAT,
-    OUT agg_cost FLOAT)
-RETURNS SETOF record
-AS 'MODULE_PATHNAME', 'euclideanTSP'
-LANGUAGE C VOLATILE STRICT;
-
--- COMMENTS
-
-COMMENT ON FUNCTION _pgr_TSPannealingEuclidean(TEXT, BIGINT, BIGINT, FLOAT, INTEGER, INTEGER, INTEGER, FLOAT, FLOAT, FLOAT, BOOLEAN)
-IS 'pgRouting internal function';
