+++ src/tsp/CMakeLists.txt
@@ -1,9 +1,7 @@
 ADD_LIBRARY(tsp OBJECT
-    euclideanTSP.c
-    TSP.c
-
-    TSP_driver.cpp
-    euclideanTSP_driver.cpp
+    TSPannealing.c
+    TSPannealing_driver.cpp
+    TSPannealingEuclidean_driver.cpp
 
     euclideanDmatrix.cpp
     tour.cpp
