+++ test/tsp/doc-pgr_TSPannealing.result
@@ -29,36 +29,36 @@ SELECT * FROM pgr_TSPannealing(
   14 |    1 |    0 |       18
 (14 rows)
 
+-- q1.1
 -- q2
 SELECT * FROM pgr_TSPannealing(
     $$
-    SELECT * FROM pgr_dijkstraCostMatrix(
-        'SELECT id, source, target, cost, reverse_cost FROM edge_table',
-        (SELECT array_agg(id) FROM edge_table_vertices_pgr WHERE id < 14),
-        directed := false
-    )
+    SELECT id, st_X(the_geom) AS x, st_Y(the_geom)AS y  FROM edge_table_vertices_pgr
     $$,
-    start_id := 7,
-    randomize := false
-);
- seq | node | cost | agg_cost
------+------+------+----------
-   1 |    7 |    1 |        0
-   2 |    8 |    1 |        1
-   3 |    5 |    1 |        2
-   4 |    2 |    1 |        3
-   5 |    1 |    2 |        4
-   6 |    3 |    1 |        6
-   7 |    4 |    1 |        7
-   8 |    9 |    1 |        8
-   9 |   12 |    1 |        9
-  10 |   11 |    1 |       10
-  11 |   10 |    1 |       11
-  12 |   13 |    3 |       12
-  13 |    6 |    3 |       15
-  14 |    7 |    0 |       18
-(14 rows)
+    randomize := false);
+ seq | node |       cost        |     agg_cost
+-----+------+-------------------+------------------
+   1 |    1 |   1.4142135623731 |                0
+   2 |    3 |                 1 |  1.4142135623731
+   3 |    4 |                 1 | 2.41421356237309
+   4 |    9 |                 1 | 3.41421356237309
+   5 |    6 |  0.58309518948453 | 4.41421356237309
+   6 |   16 | 0.860232526704263 | 4.99730875185763
+   7 |   12 |  1.11803398874989 | 5.85754127856189
+   8 |   17 |  1.11803398874989 | 6.97557526731178
+   9 |   11 |                 1 | 8.09360925606168
+  10 |   10 |               0.5 | 9.09360925606168
+  11 |   15 |               0.5 | 9.59360925606168
+  12 |   13 |  1.58113883008419 | 10.0936092560617
+  13 |   14 |  1.58113883008419 | 11.6747480861459
+  14 |    7 |                 1 | 13.2558869162301
+  15 |    8 |                 1 | 14.2558869162301
+  16 |    5 |                 1 | 15.2558869162301
+  17 |    2 |                 1 | 16.2558869162301
+  18 |    1 |                 0 | 17.2558869162301
+(18 rows)
 
+-- q2.1
 -- q3
 SELECT * FROM pgr_TSPannealing(
     $$
@@ -80,6 +80,6 @@ SELECT * FROM pgr_TSPannealing(
    6 |    5 |    0 |      5.2
 (6 rows)
 
--- q4
+-- q3.1
 ROLLBACK;
 ROLLBACK
