+++ test/tsp/test.conf
@@ -5,13 +5,11 @@
         'comment' => 'TSP tests',
         'data' => [],
         'tests' => [qw(
-            doc-pgr_TSPannealingEuclidean
+            doc-pgr_TSPannealing
             )],
         'linux' => [qw(
-            doc-pgr_TSPannealing
             )],
         'macos' => [qw(
-            doc-pgr_TSPannealing
             )],
         # windows has a reversed path which is OK
         'windows' => [qw(
@@ -19,7 +17,6 @@
             )],
         'documentation' => [qw(
             doc-pgr_TSPannealing
-            doc-pgr_TSPannealingEuclidean
             )]
     },
 );
